requirements.txt
aioredis==2.0.1
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
DateTime==5.5
dotenv==0.9.9
exceptiongroup==1.3.0
Faker==37.11.0
fastapi==0.119.1
fastapi-cache2==0.2.2
h11==0.16.0
idna==3.11
pendulum==3.1.0
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
pyodbc==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
redis==4.6.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
zope.interface==8.0.1
